+++ chatlas/types/__init__.py
@@ -4,7 +4,9 @@ from .._chat import (  # noqa: A005
     SubmitInputArgsT,
 )
 from .._content import (
+    AudioContentTypes,
     Content,
+    ContentAudio,
     ContentCitation,
     ContentDocument,
     ContentImage,
@@ -36,7 +38,9 @@ from .._turn import FinishReason
 from .._utils import MISSING, MISSING_TYPE
 
 __all__ = (
+    "AudioContentTypes",
     "Content",
+    "ContentAudio",
     "ContentCitation",
     "ContentDocument",
     "ContentImage",